GVT-3548: Separate location tracks' switch infobox into subcomponents similar to previous refactoring
