+++ ui/src/tool-panel/location-track/location-track-switch-links-infobox.tsx
@@ -1,20 +1,11 @@
-import dialogStyles from 'geoviite-design-lib/dialog/dialog.scss';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import Infobox from 'tool-panel/infobox/infobox';
-import { LayoutLocationTrack, LayoutSwitch, LayoutSwitchId } from 'track-layout/track-layout-model';
+import { LayoutLocationTrack, LayoutSwitchId } from 'track-layout/track-layout-model';
 import { useLocationTrackInfoboxExtras } from 'track-layout/track-layout-react-utils';
-import { LayoutContext, TrackMeter } from 'common/common-model';
-import { Point } from 'model/geometry';
-import { Button, ButtonSize, ButtonVariant } from 'vayla-design-lib/button/button';
-import { Dialog, DialogVariant } from 'geoviite-design-lib/dialog/dialog';
-import { detachSwitchFromLocationTrack } from 'track-layout/layout-location-track-api';
-import { updateLocationTrackChangeTime } from 'common/change-time-api';
-import * as Snackbar from 'geoviite-design-lib/snackbar/snackbar';
+import { LayoutContext } from 'common/common-model';
 import styles from './location-track-switch-links-infobox.scss';
 import { ChangeTimes } from 'common/common-slice';
-import NavigableTrackMeter from 'geoviite-design-lib/track-meter/navigable-track-meter';
-import { SwitchBadge } from 'geoviite-design-lib/switch/switch-badge';
 import { OnSelectOptions } from 'selection/selection-model';
 import { ShowMoreButton } from 'show-more-button/show-more-button';
 import { LayoutValidationIssue, ValidatedLocationTrack } from 'publication/publication-model';
@@ -25,8 +16,7 @@ import {
 import { LoaderStatus, useLoaderWithStatus } from 'utils/react-utils';
 import { getSwitches } from 'track-layout/layout-switch-api';
 import InfoboxContent from 'tool-panel/infobox/infobox-content';
-import infoboxStyles from 'tool-panel/infobox/infobox.module.scss';
-import { createClassName } from 'vayla-design-lib/utils';
+import { LocationTrackSwitchRow } from './location-track-switch-row';
 
 const maxSwitchesToDisplay = 10;
 
@@ -73,28 +63,6 @@ export const LocationTrackSwitchLinksInfobox: React.FC<
     const switches =
         useLocationTrackInfoboxExtras(locationTrack.id, layoutContext, changeTimes)[0]?.switches ??
         [];
-    const [showingDialogToDetachSwitch, setShowingDialogToDetachSwitch] = React.useState<
-        { id: LayoutSwitchId; name: string } | undefined
-    >(undefined);
-    const [isDetachingSwitch, setIsDetachingSwitch] = React.useState(false);
-
-    const detachSwitch = () => {
-        if (showingDialogToDetachSwitch === undefined) return;
-        setIsDetachingSwitch(true);
-        detachSwitchFromLocationTrack(
-            layoutContext.branch,
-            locationTrack.id,
-            showingDialogToDetachSwitch.id,
-        )
-            .then(() => updateLocationTrackChangeTime())
-            .then(() => {
-                setIsDetachingSwitch(false);
-                setShowingDialogToDetachSwitch(undefined);
-                Snackbar.success(
-                    'tool-panel.location-track.detach-switch-links-dialog.success-toast',
-                );
-            });
-    };
     const switchIds = switches.map((s) => s.switchId);
 
     const [switchItems, switchItemLoadStatus] = useLoaderWithStatus(
@@ -131,169 +99,58 @@ export const LocationTrackSwitchLinksInfobox: React.FC<
     const [showAllSwitches, setShowAllSwitches] = React.useState(false);
 
     return (
-        <>
-            <Infobox
-                title={t('tool-panel.location-track.switch-links.heading')}
-                qa-id={'location-track-switch-links-infobox'}
-                contentVisible={contentVisible}
-                onContentVisibilityChange={onContentVisibilityChange}>
-                <ProgressIndicatorWrapper
-                    indicator={ProgressIndicatorType.Area}
-                    inProgress={
-                        switchItemLoadStatus !== LoaderStatus.Ready ||
-                        validationLoaderStatus !== LoaderStatus.Ready
-                    }
-                    inline={true}>
-                    <InfoboxContent>
-                        {switchesAll.length > 0 ? (
-                            <React.Fragment>
-                                <div className={styles['location-track-switch-links-infobox-list']}>
-                                    {(showAllSwitches
-                                        ? switchesAll
-                                        : switchesAll.slice(0, maxSwitchesToDisplay)
-                                    ).map(({ switchItem, location, address, validationIssues }) => (
-                                        <LocationTrackSwitchLink
-                                            key={switchItem.id}
-                                            layoutContext={layoutContext}
-                                            switchItem={switchItem}
-                                            location={location}
-                                            displayAddress={address}
-                                            validationIssues={validationIssues}
-                                            setShowingDialogToDetachSwitch={
-                                                setShowingDialogToDetachSwitch
-                                            }
-                                            onSelect={onSelect}
-                                        />
-                                    ))}
-                                </div>
-                                {switches.length > maxSwitchesToDisplay && (
-                                    <ShowMoreButton
-                                        expanded={showAllSwitches}
-                                        onShowMore={() => setShowAllSwitches(!showAllSwitches)}
-                                        showMoreText={t(
-                                            'tool-panel.location-track.switch-links.show-more',
-                                            {
-                                                count: switchesAll.length,
-                                            },
-                                        )}
+        <Infobox
+            title={t('tool-panel.location-track.switch-links.heading')}
+            qa-id={'location-track-switch-links-infobox'}
+            contentVisible={contentVisible}
+            onContentVisibilityChange={onContentVisibilityChange}>
+            <ProgressIndicatorWrapper
+                indicator={ProgressIndicatorType.Area}
+                inProgress={
+                    switchItemLoadStatus !== LoaderStatus.Ready ||
+                    validationLoaderStatus !== LoaderStatus.Ready
+                }
+                inline={true}>
+                <InfoboxContent>
+                    {switchesAll.length > 0 ? (
+                        <React.Fragment>
+                            <div className={styles['location-track-switch-links-infobox-list']}>
+                                {(showAllSwitches
+                                    ? switchesAll
+                                    : switchesAll.slice(0, maxSwitchesToDisplay)
+                                ).map(({ switchItem, location, address, validationIssues }) => (
+                                    <LocationTrackSwitchRow
+                                        key={switchItem.id}
+                                        layoutContext={layoutContext}
+                                        switchItem={switchItem}
+                                        location={location}
+                                        displayAddress={address}
+                                        validationIssues={validationIssues}
+                                        locationTrack={locationTrack}
+                                        onSelect={onSelect}
                                     />
-                                )}
-                            </React.Fragment>
-                        ) : (
-                            <p className={'infobox__text'}>
-                                {t('tool-panel.location-track.switch-links.no-switches')}
-                            </p>
-                        )}
-                    </InfoboxContent>
-                </ProgressIndicatorWrapper>
-            </Infobox>
-            {showingDialogToDetachSwitch && (
-                <Dialog
-                    title={t('tool-panel.location-track.detach-switch-links-dialog.title')}
-                    variant={DialogVariant.DARK}
-                    allowClose={true}
-                    onClose={() => setShowingDialogToDetachSwitch(undefined)}
-                    footerContent={
-                        <>
-                            <Button
-                                onClick={() => setShowingDialogToDetachSwitch(undefined)}
-                                variant={ButtonVariant.SECONDARY}
-                                disabled={isDetachingSwitch}>
-                                {t('button.cancel')}
-                            </Button>
-                            <div className={dialogStyles['dialog__footer-content--right-aligned']}>
-                                <Button
-                                    disabled={isDetachingSwitch}
-                                    isProcessing={isDetachingSwitch}
-                                    variant={ButtonVariant.PRIMARY_WARNING}
-                                    onClick={() => detachSwitch()}>
-                                    {t(
-                                        'tool-panel.location-track.detach-switch-links-dialog.detach-button',
-                                    )}
-                                </Button>
+                                ))}
                             </div>
-                        </>
-                    }>
-                    <div className={'dialog__text'}>
-                        {t('tool-panel.location-track.detach-switch-links-dialog.message', {
-                            switchName: showingDialogToDetachSwitch.name,
-                            trackName: locationTrack.name,
-                        })}
-                    </div>
-                </Dialog>
-            )}
-        </>
-    );
-};
-
-type ShowingDialogToDetachSwitch = { id: LayoutSwitchId; name: string };
-
-type LocationTrackSwitchLinkProps = {
-    layoutContext: LayoutContext;
-    switchItem: LayoutSwitch;
-    validationIssues: LayoutValidationIssue[];
-    location?: Point;
-    displayAddress?: TrackMeter;
-    setShowingDialogToDetachSwitch: (detach: ShowingDialogToDetachSwitch) => void;
-    onSelect: (items: OnSelectOptions) => void;
-};
-const LocationTrackSwitchLink: React.FC<LocationTrackSwitchLinkProps> = ({
-    layoutContext,
-    switchItem,
-    validationIssues,
-    location,
-    displayAddress,
-    setShowingDialogToDetachSwitch,
-    onSelect,
-}) => {
-    const { t } = useTranslation();
-    const remarkClassNames = createClassName(
-        styles['location-track-switch-links-infobox-list__remark'],
-        infoboxStyles['infobox__list-cell--strong'],
-    );
-
-    return (
-        <>
-            <div
-                title={validationIssues
-                    .map((issue) => t(issue.localizationKey, issue.params))
-                    .join('\n')}>
-                <SwitchBadge
-                    switchItem={switchItem}
-                    switchIsValid={validationIssues.length === 0}
-                    onClick={() =>
-                        onSelect({
-                            switches: [switchItem.id],
-                            selectedTab: { id: switchItem.id, type: 'SWITCH' },
-                        })
-                    }
-                />
-            </div>
-            <div className={infoboxStyles['infobox__list-cell--strong']}>
-                {displayAddress === undefined ? (
-                    t('tool-panel.location-track.switch-links.no-location')
-                ) : (
-                    <NavigableTrackMeter
-                        trackMeter={displayAddress}
-                        displayDecimals={false}
-                        location={location}
-                    />
-                )}
-            </div>
-            <div className={remarkClassNames}>
-                {switchItem.stateCategory === 'NOT_EXISTING' &&
-                    t('tool-panel.location-track.switch-links.not-existing')}
-            </div>
-            <div>
-                {layoutContext.publicationState === 'DRAFT' && (
-                    <Button
-                        size={ButtonSize.SMALL}
-                        variant={ButtonVariant.GHOST}
-                        onClick={() => setShowingDialogToDetachSwitch(switchItem)}>
-                        {t('tool-panel.location-track.switch-links.detach')}
-                    </Button>
-                )}
-            </div>
-        </>
+                            {switches.length > maxSwitchesToDisplay && (
+                                <ShowMoreButton
+                                    expanded={showAllSwitches}
+                                    onShowMore={() => setShowAllSwitches(!showAllSwitches)}
+                                    showMoreText={t(
+                                        'tool-panel.location-track.switch-links.show-more',
+                                        {
+                                            count: switchesAll.length,
+                                        },
+                                    )}
+                                />
+                            )}
+                        </React.Fragment>
+                    ) : (
+                        <p className={'infobox__text'}>
+                            {t('tool-panel.location-track.switch-links.no-switches')}
+                        </p>
+                    )}
+                </InfoboxContent>
+            </ProgressIndicatorWrapper>
+        </Infobox>
     );
 };
